putting function for creation of browser output file in utils
