+++ methplotlib/utils.py
@@ -8,6 +8,7 @@ import logging
 import binascii
 import gzip
 from pyfaidx import Fasta
+import plotly
 
 
 class Region(object):
@@ -160,3 +161,34 @@ def file_sniffer(filename):
         return "nanopolish_freq"
     else:
         sys.exit("\n\n\nInput file {} not recognized!\n".format(filename))
+
+
+def create_browser_output(fig, outfile, window):
+    if outfile is None:
+        outfile = "methylation_browser_{}.html".format(window.string)
+    else:
+        from pathlib import Path
+
+        outfile = outfile.format(region=window.string)
+        p = Path(outfile)
+        Path.mkdir(p.parent, exist_ok=True, parents=True)
+
+    if outfile.endswith(".html"):
+        write_html_output(fig, outfile)
+    else:
+        try:
+            fig.write_image(outfile)
+        except ValueError as e:
+            sys.stderr.write("\n\nERROR: creating the image in this file format failed.\n")
+            sys.stderr.write("ERROR: creating in default html format instead.\n")
+            sys.stderr.write("ERROR: additional packages required. Detailed error:\n")
+            sys.stderr.write(str(e))
+            write_html_output(fig, outfile)
+
+
+def write_html_output(fig, outfile):
+    with open(outfile, "w+") as output:
+        output.write(plotly.offline.plot(fig,
+                                         output_type="div",
+                                         show_link=False,
+                                         include_plotlyjs='cdn'))
